requirements.txt
anthropic
transformers
torch
pytest
python-dotenv
openai>=1.0.0
huggingface_hub